# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MemoryMesh API — AI Meeting Memory System powered by Cognee
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List all meetings
      parameters:
        - name: project_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
    post:
      operationId: createMeeting
      tags: [meetings]
      summary: Create and process a new meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingInput"
      responses:
        "201":
          description: Meeting created and memory ingestion started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get meeting details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteMeeting
      tags: [meetings]
      summary: Delete a meeting and forget its memories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"
  /meetings/{id}/process:
    post:
      operationId: processMeeting
      tags: [meetings]
      summary: Re-trigger memory extraction for a meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessingStatus"
  /meetings/{id}/entities:
    get:
      operationId: getMeetingEntities
      tags: [meetings]
      summary: Get all extracted entities from a meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entity"

  # ─── MEMORY LIFECYCLE ─────────────────────────────────────────────────────────
...
  /graph:
    get:
      operationId: getGraph
      tags: [graph]
      summary: Get the full knowledge graph for visualization
      parameters:
        - name: project_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: entity_type
          in: query
          schema:
            type: ["string", "null"]
        - name: depth
          in: query
          schema:
            type: integer
            default: 3
      responses:
        "200":
          description: Graph data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphData"
  /graph/node/{nodeId}:
    get:
      operationId: getGraphNode
      tags: [graph]
      summary: Get node details with its connections
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Node details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphNode"

  # ─── SEARCH ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Meeting:
      type: object
      required: [id, title, date, status, participant_count, entity_count]
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, processing, indexed, failed]
        participant_count:
          type: integer
        entity_count:
          type: integer
        project_names:
          type: array
          items:
            type: string
        summary:
          type: ["string", "null"]
        duration_minutes:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    MeetingDetail:
      type: object
      required: [id, title, date, status, transcript, entities, decisions, tasks, participants]
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: string
          format: date-time
        status:
          type: string
        transcript:
          type: string
        summary:
          type: ["string", "null"]
        entities:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
        decisions:
          type: array
          items:
            $ref: "#/components/schemas/Decision"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
        risks:
          type: array
          items:
            type: string
        blockers:
          type: array
          items:
            type: string
        topics:
          type: array
          items:
            type: string
        related_meetings:
          type: array
          items:
            $ref: "#/components/schemas/Meeting"
        duration_minutes:
          type: ["integer", "null"]
    MeetingInput:
      type: object
      required: [title, date, content]
      properties:
        title:
          type: string
          minLength: 1
        date:
          type: string
          format: date-time
        content:
          type: string
          minLength: 1
          description: Meeting transcript or notes text
        content_type:
          type: string
          enum: [transcript, notes, summary]
          default: transcript
        project_id:
          type: ["integer", "null"]
        duration_minutes:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
    ProcessingStatus:
      type: object
      required: [meeting_id, status, message]
      properties:
        meeting_id:
          type: integer
        status:
          type: string
        message:
          type: string
        progress:
          type: ["integer", "null"]

    # ─── ENTITIES ──────────────────────────────────────────────────────────────
    Entity:
      type: object
      required: [id, name, type, meeting_count]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [person, project, topic, decision, task, risk, blocker, document, question, deadline]
        description:
          type: ["string", "null"]
        meeting_count:
          type: integer
        first_seen:
          type: ["string", "null"]
          format: date-time
        last_seen:
          type: ["string", "null"]
          format: date-time
        attributes:
          type: object
          additionalProperties: true
    Participant:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        meeting_count:
          type: integer
...
    Decision:
      type: object
      required: [id, title, status, meeting_id, created_at]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, revised, superseded, implemented, rejected]
        meeting_id:
          type: integer
        meeting_title:
          type: ["string", "null"]
        project_name:
          type: ["string", "null"]
        assigned_to:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        revision_count:
          type: integer
    Task:
      type: object
      required: [id, title, status, created_at]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, in_progress, completed, blocked, overdue]
        assigned_to:
          type: ["string", "null"]
        due_date:
          type: ["string", "null"]
          format: date-time
        project_name:
          type: ["string", "null"]
        meeting_id:
          type: integer
        created_at:
          type: string
          format: date-time

    # ─── MEMORY ────────────────────────────────────────────────────────────────
...
    OperationResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        affected_count:
          type: integer

    # ─── GRAPH ─────────────────────────────────────────────────────────────────
    GraphData:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNodeSummary"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
        node_count:
          type: integer
        edge_count:
          type: integer
    GraphNodeSummary:
      type: object
      required: [id, label, type, weight]
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
        weight:
          type: number
        entity_id:
          type: ["integer", "null"]
        color:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
    GraphNode:
      type: object
      required: [id, label, type, connections]
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
        entity_id:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        connections:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
        meetings:
          type: array
          items:
            $ref: "#/components/schemas/Meeting"
        attributes:
          type: object
          additionalProperties: true
    GraphEdge:
      type: object
      required: [source, target, relationship]
      properties:
        source:
          type: string
        target:
          type: string
        relationship:
          type: string
        weight:
          type: number
        meeting_id:
          type: ["integer", "null"]

    # ─── SEARCH ────────────────────────────────────────────────────────────────
...